meu_projeto/usuarios/requirements.txt
Django==5.1
djangorestframework==3.14
